+++ data/remote/src/main/kotlin/com/berlin/remote/TvShowDetailsRemoteDataSourceImpl.kt
@@ -1,6 +1,8 @@
 package com.berlin.remote
 
+import android.util.Log
 import com.berlin.repository.datasource.remote.TvShowDetailsRemoteDataSource
+import com.berlin.repository.datasource.remote.dto.MediaImagesResponse
 import com.berlin.repository.datasource.remote.dto.TVShowDetailsDto
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
@@ -10,6 +12,15 @@ import io.ktor.client.request.parameter
 class TvShowDetailsRemoteDataSourceImpl(
     private val client: HttpClient
 ) : TvShowDetailsRemoteDataSource {
+    override suspend fun getSeriesImages(id: Long): MediaImagesResponse {
+        Log.d("Khairy", "getMovieImages from remote data source...")
+        return client.get(
+            ApiConstants.SERIES_IMAGES
+                .replace("id", id.toString())
+        ) {
+        }.body<MediaImagesResponse>()
+    }
+
     override suspend fun getTvShowDetails(id: Long, language: String): TVShowDetailsDto {
         return client.get("tv/$id"){
             parameter(ApiConstants.LANGUAGE, language)
