--- data/repository/src/main/java/com/berlin/repository/datasource/remote/SeriesDetailsRemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.berlin.repository.datasource.remote
-
-import com.berlin.repository.datasource.remote.dto.MediaImagesResponse
-
-interface SeriesDetailsRemoteDataSource {
-    suspend fun getSeriesImages(id: Long): MediaImagesResponse
-}
