+++ data/repository/src/main/java/com/berlin/repository/datasource/remote/TvShowDetailsRemoteDataSource.kt
@@ -1,7 +1,9 @@
 package com.berlin.repository.datasource.remote
 
+import com.berlin.repository.datasource.remote.dto.MediaImagesResponse
 import com.berlin.repository.datasource.remote.dto.TVShowDetailsDto
 
 interface TvShowDetailsRemoteDataSource {
+    suspend fun getSeriesImages(id: Long): MediaImagesResponse
     suspend fun getTvShowDetails(id: Long,language: String): TVShowDetailsDto
 }
