--- domain/usecase/src/main/kotlin/repository/SeriesDetailsRepository.kt
@@ -1,6 +1,0 @@
-package repository
-
-interface SeriesDetailsRepository {
-    suspend fun getSeriesImages(id:Long):List<String>
-
-}
