+++ domain/usecase/src/main/kotlin/repository/TvShowDetailsRepository.kt
@@ -4,4 +4,5 @@ import com.berlin.entity.TvShowDetails
 
 interface TvShowDetailsRepository {
     suspend fun getTvShowDetails(id: Long, language: String): TvShowDetails?
+    suspend fun getSeriesImages(id:Long):List<String>
 }
