+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/MediaDetailsScreen.kt
@@ -14,6 +14,9 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.statusBarsPadding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -28,6 +31,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.buildAnnotatedString
@@ -267,6 +271,29 @@ fun ExpandableDescription(
         ),
         textAlign = TextAlign.Start
     )
+}
+
+@Composable
+fun MediaGallery(modifier: Modifier = Modifier, mediaImages: List<String>) {
+    LazyVerticalGrid(
+        modifier = modifier,
+        columns = GridCells.Adaptive(minSize = 160.dp),
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        items(mediaImages.size) { index: Int ->
+            AsyncImage(
+                model = mediaImages[index],
+                contentDescription = stringResource(com.berlin.ui.R.string.media_image),
+                modifier = Modifier
+                    .width(160.dp)
+                    .height(145.dp),
+                contentScale = ContentScale.Fit,
+//                placeholder = ,
+//                error = ,
+            )
+        }
+    }
 }
 
 @Preview(showBackground = true)
